+++ CHANGELOG.md
@@ -58,6 +58,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix permission for `checkoutCustomerAttach` mutation - #5192 by @maarcingebala
 - Restrict access to user field - #5194 by @maarcingebala
 - Unify permission for service account api client in test - #5197 by @fowczarek
+- Add additional confirmation step to checkoutComplete mutation - #5179 by @salwator
 - Allow sorting warehouses by name - #5211 by @dominik-zeglen
 - Add anonymization to GraphQL's `webhookSamplePayload` endpoint - #5161 @derenio
 - Add mutation for assigning, unassigning shipping zones to warehouse - #5217 by @kswiatek92